pyproject.toml
[project]
name = "crud-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "langchain>=1.2.10",
    "psycopg2>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt[crypto]>=2.11.0",
    "sqlmodel>=0.0.37",
]
